finished part4

+++ part4/controllers/blogs.js
@@ -2,8 +2,9 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
 
+
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)  
 })
   
@@ -14,16 +15,28 @@ blogsRouter.post('/', async (request, response) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: request.user._id
   })
   
   const savedBlog = await blog.save()
+  request.user.blogs = request.user.blogs.concat(savedBlog._id)
+  await request.user.save()
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() === request.user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    request.user.blogs = request.user.blogs.filter(b => b.toString() !== blog._id.toString())
+    await request.user.save()
+    response.status(204).end()
+  } else {
+    return response.status(400).json({ error: 'must be the creator of the blog' })
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
